UI codebase fully converted to scriptable objects

+++ Assets/Scripts/UI/Player/HPManager.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "HP Manager", menuName = "ScriptableObjects/HP Manager")]
 public class HPManager : ScriptableObject
 {
-    public int maxHP;
-    public int currentHP;
+    public float maxHP;
+    public float currentHP;
 }
